Split landfill_warning.rs into landfill_warning/ subdirectory
Split crates/simulation/src/landfill_warning.rs (659 lines) into focused modules
under landfill_warning/ for better maintainability:
- types.rs (69 lines): LandfillWarningTier, LandfillWarningEvent, LandfillCapacityState
- calculations.rs (69 lines): Pure logic functions and constants
- systems.rs (85 lines): Bevy system and plugin
- tests.rs (430 lines): All unit tests
- mod.rs (11 lines): Thin facade with re-exports
