+++ crates/simulation/src/landfill_warning/tests.rs
@@ -1,221 +1,7 @@
-use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::garbage::WasteSystem;
-use crate::services::{ServiceBuilding, ServiceType};
-use crate::SlowTickTimer;
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-/// Each Landfill service building provides this many tons of capacity.
-pub const LANDFILL_CAPACITY_PER_BUILDING: f64 = 500_000.0;
-
-/// Threshold for Low warning tier: 25% remaining capacity.
-const LOW_THRESHOLD_PCT: f64 = 25.0;
-
-/// Threshold for Critical warning tier: 10% remaining capacity.
-const CRITICAL_THRESHOLD_PCT: f64 = 10.0;
-
-/// Threshold for VeryLow warning tier: 5% remaining capacity.
-const VERY_LOW_THRESHOLD_PCT: f64 = 5.0;
-
-/// Days per year for years_remaining calculation.
-const DAYS_PER_YEAR: f32 = 365.0;
-
-// =============================================================================
-// Types
-// =============================================================================
-
-/// Warning tier indicating how much landfill capacity remains.
-///
-/// Ordered from least severe (Normal) to most severe (Emergency).
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum LandfillWarningTier {
-    /// More than 25% capacity remaining. No warnings needed.
-    #[default]
-    Normal,
-    /// 10%--25% capacity remaining. Advisory warning.
-    Low,
-    /// 5%--10% capacity remaining. Urgent warning.
-    Critical,
-    /// 0%--5% capacity remaining. Severe warning.
-    VeryLow,
-    /// 0% capacity remaining. Collection halted.
-    Emergency,
-}
-
-impl LandfillWarningTier {
-    /// Returns a human-readable label for the tier.
-    pub fn label(self) -> &'static str {
-        match self {
-            Self::Normal => "Normal",
-            Self::Low => "Low Capacity",
-            Self::Critical => "Critical",
-            Self::VeryLow => "Very Low",
-            Self::Emergency => "Emergency",
-        }
-    }
-}
-
-/// Event fired whenever the landfill warning tier changes.
-#[derive(Event, Debug, Clone)]
-pub struct LandfillWarningEvent {
-    /// The new warning tier after the change.
-    pub tier: LandfillWarningTier,
-    /// Remaining capacity as a percentage (0.0 to 100.0).
-    pub remaining_pct: f32,
-}
-
-/// City-wide landfill capacity tracking resource.
-///
-/// Updated each slow tick by `update_landfill_capacity`. Other systems can read
-/// `collection_halted` to stop waste collection when capacity is exhausted.
-#[derive(Resource, Default, Clone, Debug, Serialize, Deserialize)]
-pub struct LandfillCapacityState {
-    /// Total landfill capacity across all Landfill buildings (tons).
-    pub total_capacity: f64,
-    /// Current fill level (tons). Increases each slow tick by daily input.
-    pub current_fill: f64,
-    /// Daily waste input rate going to landfills (tons/day).
-    pub daily_input_rate: f64,
-    /// Estimated days until capacity is exhausted at current fill rate.
-    pub days_remaining: f32,
-    /// Estimated years until capacity is exhausted (days_remaining / 365).
-    pub years_remaining: f32,
-    /// Remaining capacity as a percentage (0.0 to 100.0).
-    pub remaining_pct: f32,
-    /// Current warning tier derived from remaining_pct.
-    pub current_tier: LandfillWarningTier,
-    /// When true (Emergency tier), waste collection should be halted by
-    /// downstream systems.
-    pub collection_halted: bool,
-    /// Number of Landfill service buildings in the city.
-    pub landfill_count: u32,
-}
-
-// =============================================================================
-// Pure logic (testable without ECS)
-// =============================================================================
-
-/// Determine the warning tier from a remaining-capacity percentage (0..=100).
-pub fn tier_from_remaining_pct(remaining_pct: f64) -> LandfillWarningTier {
-    if remaining_pct <= 0.0 {
-        LandfillWarningTier::Emergency
-    } else if remaining_pct <= VERY_LOW_THRESHOLD_PCT {
-        LandfillWarningTier::VeryLow
-    } else if remaining_pct <= CRITICAL_THRESHOLD_PCT {
-        LandfillWarningTier::Critical
-    } else if remaining_pct <= LOW_THRESHOLD_PCT {
-        LandfillWarningTier::Low
-    } else {
-        LandfillWarningTier::Normal
-    }
-}
-
-/// Compute remaining percentage given total capacity and current fill.
-/// Returns 100.0 when total_capacity is zero (no landfills built yet, so
-/// nothing to warn about until waste actually arrives -- handled by the
-/// system which checks for zero capacity separately).
-pub fn compute_remaining_pct(total_capacity: f64, current_fill: f64) -> f64 {
-    if total_capacity <= 0.0 {
-        return 0.0;
-    }
-    let remaining = (total_capacity - current_fill).max(0.0);
-    (remaining / total_capacity * 100.0).clamp(0.0, 100.0)
-}
-
-/// Compute days remaining given remaining capacity and daily input rate.
-/// Returns `f32::INFINITY` when daily_input_rate is zero or negative.
-pub fn compute_days_remaining(
-    total_capacity: f64,
-    current_fill: f64,
-    daily_input_rate: f64,
-) -> f32 {
-    if daily_input_rate <= 0.0 {
-        return f32::INFINITY;
-    }
-    let remaining = (total_capacity - current_fill).max(0.0);
-    (remaining / daily_input_rate) as f32
-}
-
-/// Apply one slow tick of fill accumulation. Returns new fill level,
-/// clamped to total_capacity.
-pub fn advance_fill(current_fill: f64, daily_input: f64, total_capacity: f64) -> f64 {
-    (current_fill + daily_input).min(total_capacity)
-}
-
-// =============================================================================
-// Bevy system
-// =============================================================================
-
-/// Updates landfill capacity state each slow tick.
-///
-/// 1. Counts `Landfill` service buildings to compute total capacity.
-/// 2. Reads `WasteSystem.total_generated_tons` for the daily input rate.
-/// 3. Advances fill level by one day's input per slow tick.
-/// 4. Computes remaining percentage, days/years remaining, and warning tier.
-/// 5. Fires `LandfillWarningEvent` when the tier changes.
-/// 6. Sets `collection_halted` at Emergency tier.
-pub fn update_landfill_capacity(
-    slow_timer: Res<SlowTickTimer>,
-    waste_system: Res<WasteSystem>,
-    buildings: Query<&ServiceBuilding>,
-    mut state: ResMut<LandfillCapacityState>,
-    mut warning_events: EventWriter<LandfillWarningEvent>,
-) {
-    if !slow_timer.should_run() {
-        return;
-    }
-
-    // 1. Count landfill buildings and compute total capacity.
-    let landfill_count = buildings
-        .iter()
-        .filter(|b| b.service_type == ServiceType::Landfill)
-        .count() as u32;
-    let total_capacity = landfill_count as f64 * LANDFILL_CAPACITY_PER_BUILDING;
-
-    // 2. Read daily waste input.
-    let daily_input_rate = waste_system.period_generated_tons;
-
-    // 3. Advance fill level (one slow tick ~ one game day).
-    let current_fill = advance_fill(state.current_fill, daily_input_rate, total_capacity);
-
-    // 4. Compute derived metrics.
-    let remaining_pct = compute_remaining_pct(total_capacity, current_fill);
-    let days_remaining = compute_days_remaining(total_capacity, current_fill, daily_input_rate);
-    let years_remaining = days_remaining / DAYS_PER_YEAR;
-    let new_tier = tier_from_remaining_pct(remaining_pct);
-
-    // 5. Fire event on tier change.
-    let old_tier = state.current_tier;
-    if new_tier != old_tier {
-        warning_events.send(LandfillWarningEvent {
-            tier: new_tier,
-            remaining_pct: remaining_pct as f32,
-        });
-    }
-
-    // 6. Update state.
-    state.total_capacity = total_capacity;
-    state.current_fill = current_fill;
-    state.daily_input_rate = daily_input_rate;
-    state.days_remaining = days_remaining;
-    state.years_remaining = years_remaining;
-    state.remaining_pct = remaining_pct as f32;
-    state.current_tier = new_tier;
-    state.collection_halted = new_tier == LandfillWarningTier::Emergency;
-    state.landfill_count = landfill_count;
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use super::super::calculations::*;
+    use super::super::types::*;
 
     // -------------------------------------------------------------------------
     // tier_from_remaining_pct
@@ -640,20 +426,5 @@ mod tests {
         assert_eq!(deser.current_tier, state.current_tier);
         assert_eq!(deser.collection_halted, state.collection_halted);
         assert_eq!(deser.landfill_count, state.landfill_count);
-    }
-}
-
-pub struct LandfillWarningPlugin;
-
-impl Plugin for LandfillWarningPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<LandfillCapacityState>()
-            .add_event::<LandfillWarningEvent>()
-            .add_systems(
-                FixedUpdate,
-                update_landfill_capacity
-                    .after(crate::imports_exports::process_trade)
-                    .in_set(crate::SimulationSet::Simulation),
-            );
     }
 }
